django/requirements.txt
# Django -----------------------------------------------------------------------
Django==5.0.7
django-storages==1.14.2
python-dotenv==1.0.1
# Jinja2==3.1.4
whitenoise==6.6.0
channels==4.1.0
daphne==4.1.2
django-modeltranslation==0.18.11
django-prometheus==2.3.1
polib==1.2.0
django-cors-headers==4.4.0
django-htmx-autocomplete==0.8.3
rules==3.4
django-cleanup==8.1.0
django-azure-auth==2.0.0
django-browser-reload==1.13.0

# Logging ----------------------------------------------------------------------
django-structlog[celery,commands]==8.1.0

# Database ---------------------------------------------------------------------
# mssql-django==1.5
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Azure / auth -----------------------------------------------------------------
azure-core==1.30.1
azure-identity==1.16.0
azure-keyvault==4.2.0
azure-ai-formrecognizer==3.3.3
azure-ai-translation-document==1.0.0
azure-storage-blob==12.19.1
# requests_ntlm==1.2.0

# LLM-related ------------------------------------------------------------------
openai==1.25.1
llama-index-core==0.10.33
llama-index-vector-stores-postgres==0.1.5
llama-index-llms-azure-openai==0.1.6
llama-index-embeddings-azure-openai==0.1.7
llama-index-llms-langchain==0.1.3
# llama-index-llms-groq==0.1.3
langchain-core==0.1.48
langchain-openai==0.1.3
tiktoken==0.6.0

# Document extraction ----------------------------------------------------------
beautifulsoup4==4.12.3
html2text==2024.2.26
html2docx==1.6.0
Markdown==3.6
newspaper3k==0.2.8
PyMuPDF==1.24.1
filetype==1.2.0
docxtpl==0.16.8
mammoth==1.7.1
python-pptx==0.6.23
python-docx==1.1.0

# NLP (non-LLM) ----------------------------------------------------------------
# textblob==0.18.0.post0
langdetect==1.0.9
# langcodes==3.3.0
# nltk==3.8.1
# jellyfish==1.0.3
# unidecode==1.3.8
# sumy==0.11.0
lxml==4.9.2
# pyahocorasick==2.1.0
# spacy==3.7.4

# Data science -----------------------------------------------------------------
pandas==2.2.2
shapely==2.0.3
pdf2image==1.17.0
pillow==10.3.0
pypdf==4.2.0
reportlab==4.2.0
PyPDF2==3.0.1 
pyreadstat==1.2.7
openpyxl==3.1.2
pytest


# Redis/Celery -----------------------------------------------------------------
celery==5.4.0
redis==5.0.5
django-redis==5.4.0
